Added more Doc and Tests

+++ src/entity.rs
@@ -1,4 +1,4 @@
 //! Entity type
 
 /// Entity type
-pub type Entity = u32;
+pub type Entity = u64;
